+++ src/main/java/com/simsilica/input/MovementTarget.java
@@ -1,0 +1,69 @@
+/*
+ * $Id$
+ *
+ * Copyright (c) 2020, Simsilica, LLC
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ *
+ * 2. Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in
+ *    the documentation and/or other materials provided with the
+ *    distribution.
+ *
+ * 3. Neither the name of the copyright holder nor the names of its
+ *    contributors may be used to endorse or promote products derived
+ *    from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+ * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+ * COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
+ * INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+ * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
+ * HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT,
+ * STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED
+ * OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package com.simsilica.input;
+
+import com.simsilica.mathd.*;
+
+/**
+ *
+ *
+ *  @author    Paul Speed
+ */
+public interface MovementTarget {
+
+    /**
+     *  Called once per frame to let the target update its position
+     *  in whatever implementation-specific approach that it chooses.
+     *  movementForces are the scaled vectors of force as collected
+     *  from the input devices, where z is forward, x is 'strafe', and
+     *  y is up/down.  These values are premultiplied by "currentSpeed"
+     *  but not by tpf.
+     */
+    public void move( Quatd rotation, Vec3d movementForces, double tpf );
+
+    /**
+     *  Called to set the target's rotation based on current input
+     *  state.
+     */
+    public void setRotation( Quatd rotation );
+
+    /**
+     *  Called during initialization to get the initial target rotation.
+     *  This will be used to initialize the yaw/pitch tracked by the MovementState.
+     */
+    public Quatd getRotation();
+}